fix(tools): int32 in premade_to_sgdk nearest-colour distance (int16 overflowed)
convert_image accumulated squared RGB distance in int16. A single channel delta of 255 squares to 65025, which wraps negative in int16, so distant colours could rank as nearest. Shared- and match-palette conversions silently mis-mapped far pixels - e.g. a dark wolf leg landing on a bright orb colour. Distance now accumulates in int32. test_far_color_no_overflow fails on the old code and passes on the fix; 10/10 host checks.

+++ tools/premade_to_sgdk.py
@@ -78,8 +78,10 @@ def convert_image(img, max_colors=15, palette=None):
 
     out = np.zeros((h, w), dtype=np.uint8)  # default 0 == transparent
     if palette:
-        pal = np.array(palette, dtype=np.int16)            # K,3
-        px = rgb[opaque].astype(np.int16)                  # N,3
+        # int32, not int16: a single channel delta can be 255, and 255**2 = 65025
+        # overflows int16 (wraps negative), which mis-maps far colours as "nearest".
+        pal = np.array(palette, dtype=np.int32)            # K,3
+        px = rgb[opaque].astype(np.int32)                  # N,3
         dist = ((px[:, None, :] - pal[None, :, :]) ** 2).sum(axis=2)  # N,K
         out[opaque] = (dist.argmin(axis=1) + 1).astype(np.uint8)      # 1..K
